requirements.txt
# GUI
customtkinter>=5.2.0
Pillow>=10.0.0
pystray>=0.19.5

# Device Discovery
zeroconf>=0.148.0
async-upnp-client>=0.43.0

# Device Control
tinytuya>=1.15.0
pychromecast>=14.0.0
pywizlight>=0.5.14
tapo>=0.7.0
aiohttp>=3.9.0

# Async utilities
aiofiles>=23.0.0
